requirements.txt
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
blosc2==3.2.1
build==1.2.2.post1
cachetools==5.5.2
category_encoders==2.7.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cramjam==2.9.1
cycler==0.12.1
Cython==3.0.12
dash==3.0.1
databricks-sdk==0.49.0
db-dtypes==1.4.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deltalake==0.25.4
Deprecated==1.2.18
deprecation==2.1.0
docker==7.1.0
dynaconf==3.2.10
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fastparquet==2024.11.0
Flask==3.0.3
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kedro==0.19.12
kedro-datasets==6.0.0
kedro-telemetry==0.6.2
kedro_mlflow==0.14.4
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
llvmlite==0.44.0
lxml==4.9.4
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlflow==2.21.2
mlflow-skinny==2.21.2
more-itertools==10.6.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.28.0
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
pre_commit_hooks==5.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycaret==3.3.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydata-google-auth==1.9.1
Pygments==2.19.1
pyod==2.0.4
pyodbc==5.2.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rope==1.13.0
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
Send2Trash==1.8.3
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
tables==3.10.2
tbats==1.1.3
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
wurlitzer==3.1.1
xxhash==3.5.0
yellowbrick==1.5
zipp==3.21.0
